requirements.txt
annotated-types==0.7.0
anyio==4.9.0
apturl==0.5.2
bcrypt==3.2.0
blinker==1.9.0
Brlapi==0.8.3
cachetools==5.5.2
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.2
duplicity==0.8.21
exceptiongroup==1.2.2
fasteners==0.14.1
Flask==3.1.0
future==0.18.2
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-genai==1.14.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.20.2
httpx==0.28.1
idna==3.3
importlib-metadata==4.6.4
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.6
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==3.0.2
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
paramiko==2.9.3
pexpect==4.8.0
Pillow==9.0.1
prettytable==3.16.0
proto-plus==1.26.1
protobuf==5.29.4
ptyprocess==0.7.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycairo==1.20.1
pycups==2.0.1
pydantic==2.11.4
pydantic_core==2.33.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.32.3
rsa==4.9.1
SecretStorage==3.3.1
six==1.16.0
sniffio==1.3.1
systemd-python==234
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.5
usb-creator==0.3.7
wadllib==1.3.6
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
xdg==5
xkit==0.0.0
zipp==1.0.0
